requirements.txt
asgiref==3.4.1
astroid==2.7.3
autopep8==1.5.7
certifi==2021.5.30
chardet==4.0.0
coverage==5.5
dj-database-url==0.5.0
Django==3.2.7
django-environ==0.7.0
gunicorn==20.1.0
idna==3.2
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.6
whitenoise==5.3.0
wrapt==1.12.1
django-crispy-forms==1.13.0
Pillow==8.4.0
djangorestframework==3.12.4